Move PeerList interface into p2p_types
